# fix(android): remove usage of getResourceAsStream()