--- a/.eslintignore
+++ b/.eslintignore
@@ -4,7 +4,7 @@ node_modules
 
 android/runtime/v8/tools/bootstrap.js
 android/dev
-android/modules/ui/src/java/ti/modules/titanium/ui/widget/webview/*.js
+android/modules/ui/assets/Resources/ti.internal/webview/*.js
 
 iphone/Resources/app.js