--- a/android/modules/ui/src/java/ti/modules/titanium/ui/widget/tableview/TiBaseTableViewItem.java
+++ b/android/modules/ui/src/java/ti/modules/titanium/ui/widget/tableview/TiBaseTableViewItem.java
@@ -1,6 +1,6 @@
 /**
  * Appcelerator Titanium Mobile
- * Copyright (c) 2009-2016 by Appcelerator, Inc. All Rights Reserved.
+ * Copyright (c) 2019-current by Axway, Inc. All Rights Reserved.
  * Licensed under the terms of the Apache Public License
  * Please see the LICENSE included with this distribution for details.
  */
@@ -9,25 +9,22 @@ package ti.modules.titanium.ui.widget.tableview;
 import org.appcelerator.kroll.KrollDict;
 import org.appcelerator.kroll.KrollProxy;
 import org.appcelerator.kroll.common.Log;
+import org.appcelerator.titanium.util.TiRHelper;
+import org.appcelerator.titanium.TiApplication;
 import org.appcelerator.titanium.TiC;
 import org.appcelerator.titanium.proxy.TiViewProxy;
 import org.appcelerator.titanium.util.TiConvert;
 import org.appcelerator.titanium.util.TiFileHelper;
-import org.appcelerator.titanium.util.TiPlatformHelper;
 
 import ti.modules.titanium.ui.widget.tableview.TableViewModel.Item;
 import android.app.Activity;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
+import android.content.res.Resources;
 import android.graphics.Color;
-import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.StateListDrawable;
-import android.os.Build;
 import android.os.Handler;
 import android.os.Message;
-import android.util.DisplayMetrics;
 import android.view.ViewGroup;
 
 @SuppressWarnings("deprecation")
@@ -35,8 +32,9 @@ public abstract class TiBaseTableViewItem extends ViewGroup implements Handler.C
 {
 	private static final String TAG = "TiBaseTableViewItem";
 
-	private static Bitmap childIndicatorBitmap = null;
-	private static Bitmap checkIndicatorBitmap = null;
+	private static Drawable childIndicatorDrawable = null;
+	private static Drawable checkIndicatorDrawable = null;
+	private static Resources resources = null;
 
 	protected Handler handler;
 	protected TiFileHelper tfh;
@@ -47,68 +45,23 @@ public abstract class TiBaseTableViewItem extends ViewGroup implements Handler.C
 		super(activity);
 		this.handler = new Handler(this);
 
-		if (TiBaseTableViewItem.childIndicatorBitmap == null || TiBaseTableViewItem.checkIndicatorBitmap == null) {
-			synchronized (TiBaseTableViewItem.class)
-			{
-				// recheck to so we don't leak a bitmap.
-				int density = TiPlatformHelper.applicationLogicalDensity;
-				if (childIndicatorBitmap == null) {
-					String path = "/org/appcelerator/titanium/res/drawable/btn_more_48.png"; // default medium
-
-					switch (density) {
-						case DisplayMetrics.DENSITY_HIGH:
-							path = "/org/appcelerator/titanium/res/drawable/btn_more_72.png";
-							break;
-						case DisplayMetrics.DENSITY_MEDIUM:
-							path = "/org/appcelerator/titanium/res/drawable/btn_more_48.png";
-							break;
-						case DisplayMetrics.DENSITY_LOW:
-							path = "/org/appcelerator/titanium/res/drawable/btn_more_36.png";
-							break;
-					}
-
-					if (Build.VERSION.SDK_INT >= 9 && density == DisplayMetrics.DENSITY_XHIGH) {
-						path = "/org/appcelerator/titanium/res/drawable/btn_more_96.png";
-					}
-
-					if (Build.VERSION.SDK_INT >= 16 && density >= DisplayMetrics.DENSITY_XXHIGH) {
-						path = "/org/appcelerator/titanium/res/drawable/btn_more_144.png";
-					}
-
-					if (Build.VERSION.SDK_INT >= 16 && density >= DisplayMetrics.DENSITY_XXXHIGH) {
-						path = "/org/appcelerator/titanium/res/drawable/btn_more_192.png";
-					}
-
-					childIndicatorBitmap = BitmapFactory.decodeStream(KrollDict.class.getResourceAsStream(path));
+		if (resources == null) {
+			resources = TiApplication.getInstance().getResources();
+		}
+		if (resources != null) {
+			if (childIndicatorDrawable == null) {
+				try {
+					childIndicatorDrawable = resources.getDrawable(TiRHelper.getImageRessource("drawable.btn_more"));
+				} catch (Exception e) {
+					Log.w(TAG, "Drawable 'drawable.btn_more' not found.");
 				}
-				if (checkIndicatorBitmap == null) {
-					String path =
-						"/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_48.png"; // default medium
-					switch (density) {
-						case DisplayMetrics.DENSITY_HIGH:
-							path = "/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_72.png";
-							break;
-						case DisplayMetrics.DENSITY_MEDIUM:
-							path = "/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_48.png";
-							break;
-						case DisplayMetrics.DENSITY_LOW:
-							path = "/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_36.png";
-							break;
-					}
-
-					if (Build.VERSION.SDK_INT >= 9 && density == DisplayMetrics.DENSITY_XHIGH) {
-						path = "/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_96.png";
-					}
-
-					if (Build.VERSION.SDK_INT >= 16 && density >= DisplayMetrics.DENSITY_XXHIGH) {
-						path = "/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_144.png";
-					}
-
-					if (Build.VERSION.SDK_INT >= 18 && density >= DisplayMetrics.DENSITY_XXXHIGH) {
-						path = "/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_192.png";
-					}
-
-					checkIndicatorBitmap = BitmapFactory.decodeStream(KrollDict.class.getResourceAsStream(path));
+			}
+			if (checkIndicatorDrawable == null) {
+				try {
+					checkIndicatorDrawable =
+						resources.getDrawable(TiRHelper.getImageRessource("drawable.btn_check_buttonless_on"));
+				} catch (Exception e) {
+					Log.w(TAG, "Drawable 'drawable.btn_check_buttonless_on' not found.");
 				}
 			}
 		}
@@ -137,29 +90,14 @@ public abstract class TiBaseTableViewItem extends ViewGroup implements Handler.C
 		return null;
 	}
 
-	private BitmapDrawable createDrawable(Bitmap bitmap)
+	public Drawable getHasChildDrawable()
 	{
-		try {
-			return new BitmapDrawable(bitmap);
-		} catch (Throwable t) {
-			try {
-				Log.e(TAG, t.getClass().getName() + ": " + t.getMessage(), t);
-				return null;
-			} catch (Exception e) {
-				// ignore - logging failed
-				return null;
-			}
-		}
+		return childIndicatorDrawable;
 	}
 
-	public BitmapDrawable createHasChildDrawable()
+	public Drawable getHasCheckDrawable()
 	{
-		return createDrawable(childIndicatorBitmap);
-	}
-
-	public BitmapDrawable createHasCheckDrawable()
-	{
-		return createDrawable(checkIndicatorBitmap);
+		return checkIndicatorDrawable;
 	}
 
 	public Drawable loadDrawable(String url)