--- a/android/titanium/src/java/org/appcelerator/titanium/TiApplication.java
+++ b/android/titanium/src/java/org/appcelerator/titanium/TiApplication.java
@@ -145,8 +145,6 @@ public abstract class TiApplication extends Application implements KrollApplicat
 		// Keep a reference to this application object. Accessible via static getInstance() method.
 		tiApp = this;
 
-		loadBuildProperties();
-
 		mainThreadId = Looper.getMainLooper().getThread().getId();
 
 		modules = new HashMap<String, WeakReference<KrollModule>>();
@@ -307,8 +305,7 @@ public abstract class TiApplication extends Application implements KrollApplicat
 		this.buildHash = "N/A";
 
 		// Attempt to read the "build.properties" file.
-		final String FILE_NAME = "org/appcelerator/titanium/build.properties";
-		try (InputStream stream = getClass().getClassLoader().getResourceAsStream(FILE_NAME)) {
+		try (InputStream stream = KrollAssetHelper.openAsset("Resources/ti.internal/build.properties")) {
 			if (stream != null) {
 				Properties properties = new Properties();
 				properties.load(stream);
@@ -459,6 +456,8 @@ public abstract class TiApplication extends Application implements KrollApplicat
 
 	public void postOnCreate()
 	{
+		loadBuildProperties();
+
 		KrollRuntime runtime = KrollRuntime.getInstance();
 		if (runtime != null) {
 			Log.i(TAG, "Titanium Javascript runtime: " + runtime.getRuntimeName());