… android/build/common.xml
… android/modules/ui/assets/Resources/ti.internal/webview/binding.js
… android/modules/ui/assets/Resources/ti.internal/webview/json2.js
… android/modules/ui/assets/Resources/ti.internal/webview/polling.js
… android/titanium/res/drawable/background.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_144.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_192.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_36.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_48.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_72.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_check_buttonless_on_96.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_more_144.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_more_192.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_more_36.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_more_48.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_more_72.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_more_96.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_picker_normal.9.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_picker_pressed.9.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/btn_picker_selected.9.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/default_icon.png
… android/titanium/src/java/org/appcelerator/titanium/res/drawable/photoDefault.png