@jquick-axway 2019-07-18
The below code will test our interop handling between Titanium's JavaScript and the `WebView` local HTML's JavaScript.

```javascript
var htmlText =
		'<!DOCTYPE html>' +
		'<html>' +
		'	<head>' +
		'		<meta name="viewport" content="width=device-width, initial-scale=1.0">' +
		'	</head>' +
		'	<body>' +
		'		<p>HTML JavaScript Interop Test</p>' +
		'		<p id="label"></p>' +
		'	</body>' +
		'	<script>' +
		'		document.addEventListener("DOMContentLoaded", function(e) {' +
		'			Ti.API.info("@@@ Firing events from HTML");' +
		'			Ti.App.fireEvent("app:htmlLoaded", { isRequestingStartTimer: true });' +
		'			Ti.App.fireEvent("app:htmlRequestingShowToast", { message: "DOMContentLoaded" });' +
		'		});' +
		'		function updateLabelTo(x) {' +
		'			document.getElementById("label").innerHTML = x;' +
		'		}' +
		'	</script>' +
		'</html>';

var window = Ti.UI.createWindow();
var webView = Ti.UI.createWebView({
	html: htmlText,
});
window.add(webView);
window.open();

var timerId;
var timerCount;
function updateWebView() {
	timerCount = 5;
	function onUpdate() {
		if (timerCount > 0) {
			Ti.API.info("@@@ Calling evalJS()");
			webView.evalJS("updateLabelTo('Reload in:  " + timerCount + "')");
// While async evalJS() would be better, the blocking eval tests our "binding" scripts.
//webView.evalJS("updateLabelTo('Reload in:  " + timerCount + "')", function() {});
			timerCount--;
		} else {
			Ti.API.info("@@@ Reloading WebView");
			clearInterval(timerId);
			timerId = null;
			timerCount = 5;
			webView.html = htmlText;
		}
	}
	timerId = setInterval(function(e) {
		onUpdate();
	}, 1000);
	onUpdate();
}

Ti.App.addEventListener("app:htmlLoaded", function(e) {
	Ti.API.info("@@@ Received event 'htmlLoaded'");
	if (e.isRequestingStartTimer) {
		updateWebView();
	}
});

Ti.App.addEventListener("app:htmlRequestingShowToast", function(e) {
	if (Ti.App.Android || Ti.App.Windows) {
		if (e.message) {
			Ti.UI.createNotification({
				message: e.message,
				duration: Ti.UI.NOTIFICATION_DURATION_SHORT,
			}).show();
		}
	}
});
```