@jquick-axway 2019-07-27
I just took some benchmarks.

Since every app is getting burned by us reading "build.properties" on startup, this is what I measured. I measured the old `getResourceAsStream("build.properties")` method call and this PR's `KrollAssetHelper.openAsset("build.properties")` method call that replaces it (which internally calls Google's `openAsset()` method).

**Nexus 4:**
- Before PR: `300 ms`
- After PR: `0.678 ms`
- Result: `442.5x` faster

**Pixel 2:**
- Before PR: `16 ms`
- After PR: `0.180 ms`
- Result: `88x` faster

It's so fast, I had to time it nanoseconds to measure it properly. :)

The key thing here is this PR shaves off `299.4ms` from the app startup time on our Nexus 4 (our slowest device). Gary's snapshot PR shaves off an additional `3057ms` from the Nexus 4 startup time. Every bit helps. This is looking great and our launch times now feel on-par with Google's own major apps on the same device.