@build 2019-10-15
<!--
  1 failure:  Tests have failed...
  0 warning: 
  4 messages
  1 markdown notices
  DangerID: danger-id-Danger;
-->

<table>
  <thead>
    <tr>
      <th width="50"></th>
      <th width="100%" data-danger-table="true">Fails</th>
    </tr>
  </thead>
  <tbody><tr>
      <td>:no_entry_sign:</td>
      <td>Tests have failed, see below for more information.</td>
    </tr>
  </tbody>
</table>



<table>
  <thead>
    <tr>
      <th width="50"></th>
      <th width="100%" data-danger-table="true">Messages</th>
    </tr>
  </thead>
  <tbody><tr>
      <td>:book:</td>
      <td>:thumbsup: Hey!, You deleted more code than you added. That's awesome!</td>
    </tr>
  
<tr>
      <td>:book:</td>
      <td>

  :floppy_disk: <a href="https://jenkins.appcelerator.org/job/titanium-sdk/job/titanium_mobile/job/PR-11058/39/artifact/dist/mobilesdk-8.3.0.v20191015155851-osx.zip">Here's the generated SDK zipfile</a>.
  </td>
    </tr>
  
<tr>
      <td>:book:</td>
      <td>:x: 2 tests have failed
There are 2 tests failing and 479 skipped out of 4828 total tests.</td>
    </tr>
  
<tr>
      <td>:book:</td>
      <td>:fist: The commits in this PR match our conventions! Feel free to Rebase and Merge this PR when ready.</td>
    </tr>
  </tbody>
</table>

### Tests:

<table><tr><th>Classname</th><th>Name</th><th>Time</th><th>Error</th></tr>
<tr><td>android.emulator.Titanium.UI.WebView</td><td>html-script-tag</td><td>10.001</td><td><pre>Error: timeout of 10000ms exceeded
at Titanium.&lt;anonymous&gt; (/ti-mocha.js:4290:23)</pre></td></tr>
<tr><td>android.emulator.Titanium.UI.WebView</td><td>#evalJS(string, function) - async variant</td><td>0.29</td><td><pre>Error: expected 'null' to equal '"WebView.evalJS.TEST"'
at Assertion.fail (/node_modules/should/cjs/should.js:275:17)
      at Assertion.value (/node_modules/should/cjs/should.js:356:19)
      at WebView.&lt;anonymous&gt; (/ti.ui.webview.test.js:410:31)</pre></td></tr>
</table>

<p align="right">
  Generated by :no_entry_sign: <a href="https://danger.systems/js">dangerJS</a> against 46d8f733fdbac37a73ad189dace43c63df9fc7f5
</p>