@ssekhri 2019-10-21
@garymathews there is an issue with the PR related to tableviewRow height. Any tableview row with icons is shown with more height and also the icons are displayed quite big. Use the test code from Test4 in this PR.
Tested on:
Mac OS: 10.14.5
SDK: 8.3.0.v20191016113141
Appc CLI: 7.1.1
JDK: 1.8.0_162
Node: 10.5.0
Device: Nexus 4
![Nexus4_TableView](https://user-images.githubusercontent.com/6531858/67243359-3b1c3500-f40c-11e9-84f9-af6a7e32d741.png)