@ssekhri 2019-10-28
FR Passed.
Verified on:
Mac OS: 10.14.5
SDK: 8.3.0.v20191025142840
Appc CLI: 7.1.1
JDK: 1.8.0_162
Node: 10.5.0
Studio: 5.1.4.201909061933
Device: Nexus4(v5.1.1) device, Pixel3(v10.0) emulator